tech(TTB-060): Moved Button component into individual directory
