+++ client/src/home/HomeButtons.jsx
@@ -1,4 +1,4 @@
-import Button from '../components/Button';
+import Button from '../components/Button/Button';
 
 function HomeButtons() {
     return(
